refactor(#12): restructure tracking module with Kalman and Simple trackers
